Minimax computer version: heuristic int to double

+++ minimax/heu1.h
@@ -3,7 +3,7 @@
 
 string teamName1 = "team1";
 
-int heuristic1(const State& s) {
+double heuristic1(const State& s) {
     return countResult(s);
 }
