+++ minimax/heu3.h
@@ -3,7 +3,7 @@
 
 string teamName3 = "team3";
 
-int heuristic3(const State& s) {
+double heuristic3(const State& s) {
     return heuristic(s);
 }
