+++ minimax/heu8.h
@@ -3,7 +3,7 @@
 
 string teamName8 = "team8";
 
-int heuristic8(const State& s) {
+double heuristic8(const State& s) {
     return heuristic(s);
 }
